sooda/requirements.txt
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
tzdata==2022.7
